+++ PACECloudSDK/AppKit/AppKit+Delegate.swift
@@ -19,14 +19,15 @@ public protocol AppKitDelegate: AnyObject {
     func didExitGeofence(with id: String)
     func didFailToMonitorRegion(_ region: CLRegion, error: Error)
 
-    func getAccessToken(reason: AppKit.GetAccessTokenReason, oldToken: String?, completion: @escaping ((AppKit.GetAccessTokenResponse) -> Void))
+    func getAccessToken(reason: AppKit.GetAccessTokenReason, oldToken: String?, completion: @escaping ((API.Communication.GetAccessTokenResponse) -> Void))
     func logout(completion: @escaping ((AppKit.LogoutResponse) -> Void))
     func didReceiveImageData(_ image: UIImage)
 
     func paymentRequestMerchantIdentifier(completion: @escaping (String) -> Void)
-    func didCreateApplePayPaymentRequest(_ request: PKPaymentRequest, completion: @escaping (([String: Any]?) -> Void))
+    func didCreateApplePayPaymentRequest(_ request: PKPaymentRequest, completion: @escaping (API.Communication.ApplePayRequestResponse?) -> Void)
 
     func didRequestLocationVerification(location: CLLocation, threshold: Double, completion: @escaping ((Bool) -> Void))
+    func currentLocation(completion: @escaping (CLLocation?) -> Void)
 
     func setUserProperty(key: String, value: String, update: Bool)
     func logEvent(key: String, parameters: [String: Any])
@@ -39,11 +40,12 @@ public extension AppKitDelegate {
     func didEnterGeofence(with id: String) {}
     func didExitGeofence(with id: String) {}
     func didFailToMonitorRegion(_ region: CLRegion, error: Error) {}
-    func getAccessToken(reason: AppKit.GetAccessTokenReason, oldToken: String?, completion: @escaping ((AppKit.GetAccessTokenResponse) -> Void)) {}
+    func getAccessToken(reason: AppKit.GetAccessTokenReason, oldToken: String?, completion: @escaping ((API.Communication.GetAccessTokenResponse) -> Void)) {}
     func didReceiveImageData(_ image: UIImage) {}
     func paymentRequestMerchantIdentifier(completion: @escaping (String) -> Void) { completion("") }
-    func didCreateApplePayPaymentRequest(_ request: PKPaymentRequest, completion: @escaping (([String: Any]?) -> Void)) { completion(nil) }
+    func didCreateApplePayPaymentRequest(_ request: PKPaymentRequest, completion: @escaping (API.Communication.ApplePayRequestResponse?) -> Void) { completion(nil) }
     func didRequestLocationVerification(location: CLLocation, threshold: Double, completion: @escaping ((Bool) -> Void)) { completion(false) }
+    func currentLocation(completion: @escaping (CLLocation?) -> Void) { completion(nil) }
     func setUserProperty(key: String, value: String, update: Bool) {}
     func logEvent(key: String, parameters: [String: Any]) {}
     func getConfig(key: String, completion: @escaping ((Any?) -> Void)) { completion(nil) }
@@ -87,7 +89,7 @@ extension AppKit {
         }
     }
 
-    func notifyGetAccessToken(reason: AppKit.GetAccessTokenReason, oldToken: String?, callback: @escaping ((GetAccessTokenResponse) -> Void)) {
+    func notifyGetAccessToken(reason: AppKit.GetAccessTokenReason, oldToken: String?, callback: @escaping ((API.Communication.GetAccessTokenResponse) -> Void)) {
         notifyClient { [weak self] in
             self?.delegate?.getAccessToken(reason: reason, oldToken: oldToken, completion: callback)
         }
@@ -111,7 +113,7 @@ extension AppKit {
         }
     }
 
-    func notifyApplePayRequest(with request: PKPaymentRequest, callback: @escaping (([String: Any]?) -> Void)) {
+    func notifyApplePayRequest(with request: PKPaymentRequest, callback: @escaping (API.Communication.ApplePayRequestResponse?) -> Void) {
         notifyClient { [weak self] in
             self?.delegate?.didCreateApplePayPaymentRequest(request, completion: callback)
         }
@@ -120,6 +122,12 @@ extension AppKit {
     func notifyDidRequestLocationVerfication(location: CLLocation, threshold: Double, callback: @escaping ((Bool) -> Void)) {
         notifyClient { [weak self] in
             self?.delegate?.didRequestLocationVerification(location: location, threshold: threshold, completion: callback)
+        }
+    }
+
+    func notifyCurrentLocation(callback: @escaping (CLLocation?) -> Void) {
+        notifyClient { [weak self] in
+            self?.delegate?.currentLocation(completion: callback)
         }
     }